move identity transform assertion logic to utility
